system manager added

+++ api/src/Controller/PreferenceController.php
@@ -20,6 +20,14 @@ class PreferenceController {
         return $preferences;
     }
 
+    public static function GetAllPreferencesFromSystem() {
+        global $em;
+
+        $preferences = $em->getRepository(Preference::class)->findBy(['isFromSystem' => true]);
+
+        return $preferences;
+    }
+
     public static function SavePreference(Preference $preference) {
         global $em;
 
@@ -29,15 +37,24 @@ class PreferenceController {
         return $preference;
     }
 
-    public static function DeletePreferenceBySlug(string $slug) {
+    public static function DeletePreferenceBySlug(string $slug, bool $systemOperations = false) {
         global $em;
 
         $preference = $em->getRepository(Preference::class)->findOneBy(['slug' => $slug]);
 
-        $em->remove($preference);
-        $em->flush();
-
-        return true;
+        if(!$systemOperations) {
+            if(!$preference->isFromSystem) {
+                $em->remove($preference);
+                $em->flush();
+                return true;
+            } else {
+                return false;
+            }
+        } else {
+            $em->remove($preference);
+            $em->flush();
+            return true;
+        }
     }
 
     public static function EditPreferenceBySlug(string $slug, string $value): bool {
